move "keep as-is" optimizers
* from eval/tensor/dense to eval/instruction
* minimal changes to track move in this commit
