+++ eval/CMakeLists.txt
@@ -46,28 +46,28 @@ vespa_define_module(
     src/tests/instruction/generic_peek
     src/tests/instruction/generic_reduce
     src/tests/instruction/generic_rename
+    src/tests/instruction/dense_dot_product_function
+    src/tests/instruction/dense_matmul_function
+    src/tests/instruction/dense_multi_matmul_function
+    src/tests/instruction/dense_simple_expand_function
+    src/tests/instruction/dense_tensor_peek_function
+    src/tests/instruction/index_lookup_table
     src/tests/tensor/default_value_builder_factory
     src/tests/tensor/dense_add_dimension_optimizer
     src/tests/tensor/dense_dimension_combiner
-    src/tests/tensor/dense_dot_product_function
     src/tests/tensor/dense_fast_rename_optimizer
     src/tests/tensor/dense_generic_join
     src/tests/tensor/dense_inplace_join_function
-    src/tests/tensor/dense_matmul_function
-    src/tests/tensor/dense_multi_matmul_function
     src/tests/tensor/dense_number_join_function
     src/tests/tensor/dense_pow_as_map_optimizer
     src/tests/tensor/dense_remove_dimension_optimizer
     src/tests/tensor/dense_replace_type_function
-    src/tests/tensor/dense_simple_expand_function
     src/tests/tensor/dense_simple_join_function
     src/tests/tensor/dense_simple_map_function
     src/tests/tensor/dense_single_reduce_function
     src/tests/tensor/dense_tensor_create_function
-    src/tests/tensor/dense_tensor_peek_function
     src/tests/tensor/direct_dense_tensor_builder
     src/tests/tensor/direct_sparse_tensor_builder
-    src/tests/tensor/index_lookup_table
     src/tests/tensor/instruction_benchmark
     src/tests/tensor/onnx_wrapper
     src/tests/tensor/packed_mappings
