+++ eval/src/vespa/eval/instruction/dense_matmul_function.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include <vespa/eval/eval/tensor_function.h>
-#include "dense_tensor_view.h"
+#include <vespa/eval/tensor/dense/dense_tensor_view.h>
 
 namespace vespalib::tensor {
