+++ eval/src/vespa/eval/instruction/dense_multi_matmul_function.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "dense_tensor_view.h"
+#include <vespa/eval/tensor/dense/dense_tensor_view.h>
 #include <vespa/eval/eval/tensor_function.h>
 
 namespace vespalib::tensor {
