+++ eval/src/vespa/eval/instruction/dense_simple_expand_function.h
