+++ eval/src/vespa/eval/tensor/default_tensor_engine.cpp
@@ -8,24 +8,24 @@
 #include "sparse/direct_sparse_tensor_builder.h"
 #include "dense/dense_tensor.h"
 #include "dense/typed_dense_tensor_builder.h"
-#include "dense/dense_dot_product_function.h"
+#include <vespa/eval/instruction/dense_dot_product_function.h>
 #include <vespa/eval/instruction/dense_xw_product_function.h>
-#include "dense/dense_matmul_function.h"
-#include "dense/dense_multi_matmul_function.h"
+#include <vespa/eval/instruction/dense_matmul_function.h>
+#include <vespa/eval/instruction/dense_multi_matmul_function.h>
 #include "dense/dense_fast_rename_optimizer.h"
 #include "dense/dense_add_dimension_optimizer.h"
 #include "dense/dense_single_reduce_function.h"
 #include "dense/dense_remove_dimension_optimizer.h"
-#include "dense/dense_lambda_peek_optimizer.h"
+#include <vespa/eval/instruction/dense_lambda_peek_optimizer.h>
 #include "dense/dense_lambda_function.h"
-#include "dense/dense_simple_expand_function.h"
+#include <vespa/eval/instruction/dense_simple_expand_function.h>
 #include "dense/dense_simple_join_function.h"
 #include "dense/dense_number_join_function.h"
 #include "dense/dense_pow_as_map_optimizer.h"
 #include "dense/dense_simple_map_function.h"
 #include "dense/vector_from_doubles_function.h"
 #include "dense/dense_tensor_create_function.h"
-#include "dense/dense_tensor_peek_function.h"
+#include <vespa/eval/instruction/dense_tensor_peek_function.h>
 #include <vespa/eval/eval/value.h>
 #include <vespa/eval/eval/tensor_spec.h>
 #include <vespa/eval/eval/simple_value.h>
