+++ eval/src/vespa/eval/tensor/dense/CMakeLists.txt
@@ -4,18 +4,12 @@ vespa_add_library(eval_tensor_dense OBJECT
     dense_add_dimension_optimizer.cpp
     dense_cell_range_function.cpp
     dense_dimension_combiner.cpp
-    dense_dot_product_function.cpp
     dense_fast_rename_optimizer.cpp
     dense_lambda_function.cpp
-    dense_lambda_peek_function.cpp
-    dense_lambda_peek_optimizer.cpp
-    dense_matmul_function.cpp
-    dense_multi_matmul_function.cpp
     dense_number_join_function.cpp
     dense_pow_as_map_optimizer.cpp
     dense_remove_dimension_optimizer.cpp
     dense_replace_type_function.cpp
-    dense_simple_expand_function.cpp
     dense_simple_join_function.cpp
     dense_simple_map_function.cpp
     dense_single_reduce_function.cpp
@@ -24,11 +18,9 @@ vespa_add_library(eval_tensor_dense OBJECT
     dense_tensor_cells_iterator.cpp
     dense_tensor_create_function.cpp
     dense_tensor_modify.cpp
-    dense_tensor_peek_function.cpp
     dense_tensor_reduce.cpp
     dense_tensor_value_builder.cpp
     dense_tensor_view.cpp
-    index_lookup_table.cpp
     mutable_dense_tensor_view.cpp
     onnx_wrapper.cpp
     typed_cells_dispatch.cpp
